# api/openapi.yaml
openapi: 3.0.3
info:
  title: Enzyme API
  description: REST API for Enzyme, a self-hostable Slack alternative
  version: 1.0.0
servers:
  - url: /api
    description: API server
...
paths:
...
  /workspaces/{wid}/events:
    get:
      tags: [sse]
      summary: SSE event stream
      operationId: events
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/workspaceId'
      responses:
        '200':
          description: SSE event stream
          content:
            text/event-stream:
              schema:
                type: string
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    workspaceId:
      name: wid
      in: path
      required: true
      schema:
        type: string
      description: Workspace ID
...